organized more cmds
