+++ src/main/java/frc/robot/Robot.java
@@ -17,6 +17,7 @@ import frc.robot.subsystems.*;
 import frc.robot.commands.*;
 import frc.robot.commands.auto_commands.*;
 import frc.robot.commands.control_panel.ControlPanelMotorStop;
+import frc.robot.commands.drive.DriveCommand;
 import frc.robot.commands.hopper.HopperMotorStop;
 import frc.robot.commands.shooter.ShooterStop;
 /**
