+++ src/main/java/frc/robot/commands/drive/DriveCommand.java
@@ -5,7 +5,7 @@
 // /* the project.                                                               */
 // /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.drive;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Robot;
